% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPASS.R
\name{CompPASS}
\alias{CompPASS}
\title{CompPASS}
\usage{
CompPASS(raw_dat)
}
\arguments{
\item{raw_dat}{A table of Run_id, Bait, Prey, and Spectral counts.}
}
\value{
A data frame containing the following columns:
        Bait, Prey, AvePSM, Entropy, Z, S, D, WD scores
}
\description{
CompPASS: Calculates Z-Scores, WD scores, and entropy for a given input
table of AP-MS runs
}
\author{
Qingzhou Zhang
}
