% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HG.R
\name{HG}
\alias{HG}
\title{HG}
\usage{
HG(datInput)
}
\arguments{
\item{datInput}{A table of Run_id, Prey, Spectral counts, Protein Length.}
}
\value{
A data frame containing the following columns:
        InteractorA, InteractorB, HG scores
}
\description{
HG: Calculates HG scoresa given input table of AP-MS runs
}
\author{
Qingzhou Zhang
}
